requirements.txt
Django
psycopg2-binary
python-dotenv
python-docx
PyPDF2
pdfplumber
python-magic-bin
pandas
docx-mailmerge
weasyprint
reportlab
openpyxl
django-tailwind
django-browser-reload
html2docx
mammoth